Updated Encounter Calculator
Added the ability to specify a default player level and monster challenge rating that new entities will be created with.
The user no longer selects a difficulty, all four difficulties are calculated and compared against.
Calculations now happen as values change.
Added error messages for when player levels or monster crs are invalid.